feat(modeling): moved OrthonormalFormula to maths/utils
