+++ packages/modeling/src/maths/utils/OrthonormalFormula.d.ts
@@ -1,0 +1,3 @@
+import { Plane } from '../plane/type'
+
+export class OrthonormalFormula(plane: Plane)
