+++ packages/modeling/src/maths/utils/index.d.ts
@@ -4,5 +4,6 @@ export { interpolateBetween2DPointsForY } from './interpolateBetween2DPointsForY
 export { intersect } from './intersect'
 export { solve2Linear } from './solve2Linear'
 export { sin, cos } from './trigonometry'
+export { OrthonormalFormula } from './OrthonormalFormula'
 
 export as namespace utils
