+++ packages/modeling/src/maths/utils/index.js
@@ -11,3 +11,4 @@ export { area } from './area.js'
 export { interpolateBetween2DPointsForY } from './interpolateBetween2DPointsForY.js'
 export { intersect } from './intersect.js'
 export { solve2Linear } from './solve2Linear.js'
+export { OrthonormalFormula } from './OrthonormalFormula.js'
